Refactor application entry structure
