+++ src/app/App.tsx
@@ -1,14 +1,14 @@
 import { useEffect } from "react";
-import Navbar from "./components/Navbar";
-import Footer from "./components/Footer";
+import Navbar from "../components/Navbar";
+import Footer from "../components/Footer";
 
-import Hero from "./sections/Hero";
-import About from "./sections/About";
-import Experience from "./sections/Experience";
-import Skills from "./sections/Skills";
-import Contact from "./sections/Contact";
+import Hero from "../sections/Hero";
+import About from "../sections/About";
+import Experience from "../sections/Experience";
+import Skills from "../sections/Skills";
+import Contact from "../sections/Contact";
 
-import "./App.css";
+import "../App.css";
 
 function App() {
   useEffect(() => {
